Change deployment to gh-pages
The new deploy.sh script will build and deploy the
latest version to gh-pages.

This effectively fixes #1 since now the binary, generated PDF
files are not directly tracked in the master branch.
Automated generation via Travis or Actions is not worth the
trouble right now; running deploy.sh from the command line
when it's time to deploy is fit for purpose.